requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
streamlit
